pyproject.toml
[project]
name = "ai-twitter-scanner"
version = "0.1.0"
description = "AI-powered Twitter/X Financial Sentiment & Signal Monitor"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "openai>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "chromadb>=0.4.24",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "telegram-send>=0.35",
    "python-telegram-bot>=20.8",
    "pandas>=2.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.3",
    "loguru>=0.7.2",
    "tiktoken>=0.5.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "selenium>=4.16.0",
    "webdriver-manager>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
